+++ tests/test_system.cc
@@ -31,8 +31,8 @@ TEST_CASE("system::add_particle - extends existing attribute array")
     md::attribute_key<int, struct test_attribute_a_tag> test_attribute_a = {};
     md::attribute_key<int, struct test_attribute_b_tag> test_attribute_b = {};
 
-    system.require(test_attribute_a);
-    system.require(test_attribute_b);
+    system.add_attribute(test_attribute_a);
+    system.add_attribute(test_attribute_b);
 
     CHECK(system.view(test_attribute_a).size() == 0);
     CHECK(system.view(test_attribute_b).size() == 0);
@@ -198,7 +198,7 @@ TEST_CASE("system::view_velocities - returns velocity attribute")
     }
 }
 
-TEST_CASE("system::require - creates an attribute if it does not exist")
+TEST_CASE("system::add_attribute - creates an attribute if it does not exist")
 {
     // Add 3 particles
     md::system system;
@@ -210,7 +210,7 @@ TEST_CASE("system::require - creates an attribute if it does not exist")
     md::attribute_key<int, struct test_attribute_tag> test_attribute = {};
 
     // This call creates an attribute with 3 entries
-    system.require(test_attribute);
+    system.add_attribute(test_attribute);
 
     md::array_view<int> values = system.view(test_attribute);
     CHECK(values.size() == system.particle_count());
@@ -218,13 +218,13 @@ TEST_CASE("system::require - creates an attribute if it does not exist")
     CHECK(values[1] == 0);
     CHECK(values[2] == 0);
 
-    // Update attribute values and check that an extra require() call does not
-    // tamper with the existing attribute.
+    // Update attribute values and check that an extra add_attribute() call does
+    // not tamper with the existing attribute.
     values[0] = 42;
     values[1] = 42;
     values[2] = 42;
 
-    system.require(test_attribute);
+    system.add_attribute(test_attribute);
 
     CHECK(values[0] == 42);
     CHECK(values[1] == 42);